Split extension data into separate files.
